enabled CLI APIs
