+++ QEfficient/__init__.py
@@ -7,7 +7,7 @@
 
 from QEfficient.exporter.export_hf_to_cloud_ai_100 import qualcomm_efficient_converter
 from QEfficient.loader import QEFFAutoModel  # noqa: F401
-from QEfficient.transformers.modeling_utils import transform  # noqa: F401
+from QEfficient.transformers.transform import transform  # noqa: F401
 
 # Users can use QEfficient.export for exporting models to ONNX
 export = qualcomm_efficient_converter
