+++ QEfficient/exporter/export_hf_to_cloud_ai_100.py
@@ -12,18 +12,17 @@ from typing import Optional, Tuple, Type, Union
 import torch
 from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
-import QEfficient.transformers.modeling_utils
+import QEfficient
 from QEfficient.exporter.export_utils import export_onnx, fix_onnx_fp16, generate_input_files, run_model_on_ort
-from QEfficient.loader.loader import QEFFAutoModel
+from QEfficient.loader.loader import AUTO_MODEL_MAP_TO_MODEL_TYPE_MAP, QEFFAutoModel
 from QEfficient.loader.loader_factory import (
-    AUTO_MODEL_MAP_TO_MODEL_TYPE_MAP,
     QEFF_MODEL_TYPE,
     QEFFAutoModelForCausalLM,
     QEFFBaseModel,
 )
+from QEfficient.utils._utils import load_hf_tokenizer
 from QEfficient.utils.constants import QEFF_MODELS_DIR, Constants
 from QEfficient.utils.logging_utils import logger
-from QEfficient.utils._utils import load_hf_tokenizer
 
 
 def convert_to_cloud_bertstyle(
@@ -442,7 +441,7 @@ def export_lm_model_for_cloud(model_name:str, qeff_model: QEFFAutoModelForCausal
 
 def qualcomm_efficient_converter(
     model_name: str,
-    model_kv: Optional[Type[QEFFBaseModel]] = None, # type: ignore
+    model_kv: QEFFBaseModel = None, # type: ignore
     tokenizer: Optional[Union[PreTrainedTokenizer, PreTrainedTokenizerFast]]=None,
     onnx_dir_path: Optional[str]=None,
     hf_token: Optional[str] = None,
@@ -480,7 +479,7 @@ def qualcomm_efficient_converter(
     if model_kv.is_transformed and not kv:
         raise AttributeError("Transformed model is passed while requsting to convert non-transformed model")
     
-    model_kv: Type[QEFFBaseModel] = QEfficient.transform(model_kv) if kv else model_kv
+    model_kv = model_kv if model_kv.is_transformed else QEfficient.transform(model_kv) if kv else model_kv
 
 
     if onnx_dir_path is None:
