+++ QEfficient/loader/loader.py
@@ -6,14 +6,42 @@
 # -----------------------------------------------------------------------------
 
 import os
-from typing import Any, Type
+from typing import Any, Dict, Type
 
-from QEfficient.loader.loader_factory import (
-    MODEL_TYPE_TO_QEFF_AUTO_MODEL_MAP,
-    QEFFBaseModel,
-    get_hf_model_type,
-)
+from transformers import AutoConfig
+from transformers.models.auto.modeling_auto import MODEL_FOR_CAUSAL_LM_MAPPING
+
+from QEfficient.loader.loader_factory import QEFF_MODEL_TYPE, QEFFAutoModelForCausalLM, QEFFBaseModel
 from QEfficient.utils._utils import login_and_download_hf_lm
+
+MODEL_TYPE_TO_QEFF_AUTO_MODEL_MAP: Dict[QEFF_MODEL_TYPE, Type[QEFFBaseModel]] = {
+    QEFF_MODEL_TYPE.CAUSALLM: QEFFAutoModelForCausalLM
+}
+
+AUTO_MODEL_MAP_TO_MODEL_TYPE_MAP: Dict[Type[QEFFBaseModel], QEFF_MODEL_TYPE] = {v:k for k,v in MODEL_TYPE_TO_QEFF_AUTO_MODEL_MAP.items()}
+
+def get_hf_model_type(hf_model_path: str) -> QEFF_MODEL_TYPE:
+    """
+    Loads model config file and returns the type of the model (i.e. LLMs, SD, quantized etc.) as supported by the library.
+    """
+    assert os.path.isdir(hf_model_path), "Pleae pass local dir path where the model is downloaded; use `QEfficient.utils.login_and_download_hf_lm` for downloading hf model"
+    config, kwargs = AutoConfig.from_pretrained(
+                hf_model_path,
+                return_unused_kwargs=True,
+            )
+
+    if config.__class__ in MODEL_FOR_CAUSAL_LM_MAPPING:
+        # FIXME: Add logic to handle if quantization config is stored in separate quant_config.json outside of config, also create a separate function for this and below lines
+        quant_config = getattr(config, "quantization_config", getattr(config, "quant_config", None))
+        if quant_config is not None:
+            if quant_config.get("quant_method", None) == "awq":
+                return QEFF_MODEL_TYPE.AWQ
+            else:
+                raise NotImplementedError(f"current model type is not yet supported {type(config)}")
+        else:
+            return QEFF_MODEL_TYPE.CAUSALLM
+    else:
+        raise NotImplementedError(f"model type {type(config)} is not yet supported")
 
 
 class QEFFAutoModel:
@@ -27,7 +55,7 @@ class QEFFAutoModel:
             f"using the `{self.__class__.__name__}.from_pretrained(pretrained_model_name_or_path)`")
     
     @classmethod
-    def from_pretrained(cls, pretrained_model_name_or_path: str, *args, **kwargs) -> Type[QEFFBaseModel]:
+    def from_pretrained(cls, pretrained_model_name_or_path: str, *args, **kwargs) -> QEFFBaseModel:
         """
         Downloads HuggingFace model if already doesn't exist locally, returns QEffAutoModel object based on type of model.
         """
