+++ QEfficient/loader/loader_factory.py
@@ -5,17 +5,22 @@
 #
 # ----------------------------------------------------------------------------
 
-import os
 from abc import ABC, abstractmethod
 from enum import Enum
-from typing import Any, Dict, Type
+from typing import Any
 
 import torch.nn as nn
-from transformers import AutoConfig, AutoModelForCausalLM
+from transformers import AutoModelForCausalLM
 from transformers.models.auto.modeling_auto import MODEL_FOR_CAUSAL_LM_MAPPING
 
-import QEfficient.transformers.modeling_utils
+import QEfficient
 from QEfficient.transformers.modeling_utils import TransformersToQEffModulesDict
+
+
+class QEFF_MODEL_TYPE(Enum):
+    CAUSALLM = "LLM"
+    DIFFUSION = "STABLE_DIFFUSION"
+    AWQ = "AWQ"
 
 
 class QEFFBaseModel(ABC):
@@ -29,11 +34,11 @@ class QEFFBaseModel(ABC):
         super().__init__()
         # Users can call generate or execute
         self.generate = self.execute
-    
+
     @classmethod
     def from_pretrained(cls, pretrained_model_name_or_path: str, *args, **kwargs):
         raise NotImplementedError("Must implement for child classes")
-    
+
     @property
     def is_transformed(self) -> bool:
         raise NotImplementedError("Must implement for child classes")
@@ -49,11 +54,11 @@ class QEFFBaseModel(ABC):
     @abstractmethod
     def execute(self, *args, **kwargs) -> Any:
         pass
-    
+
     @abstractmethod
     def transform(self, *args, **kwargs) -> Any:
         pass
-    
+
     @abstractmethod
     def export(self, *args, **kwargs) -> Any:
         pass
@@ -61,7 +66,7 @@ class QEFFBaseModel(ABC):
     @abstractmethod
     def compile(self, *args, **kwargs) -> Any:
         pass
-    
+
 
 class QEFFAutoModelForCausalLM(QEFFBaseModel):
     """
@@ -95,7 +100,7 @@ class QEFFAutoModelForCausalLM(QEFFBaseModel):
         raise NotImplementedError("Reached too far!!")
     
     def transform(self):
-        QEfficient.transformers.modeling_utils.transform_lm(self.model)
+        QEfficient.transform(self)
         return self
 
     def export(self):
@@ -103,40 +108,3 @@ class QEFFAutoModelForCausalLM(QEFFBaseModel):
     
     def compile(self, *args, **kwargs) -> Any:
         raise NotImplementedError("Reached too far!!")
-
-
-class QEFF_MODEL_TYPE(Enum):
-    CAUSALLM = "LLM"
-    DIFFUSION = "STABLE_DIFFUSION"
-    AWQ = "AWQ"
-
-
-MODEL_TYPE_TO_QEFF_AUTO_MODEL_MAP: Dict[QEFF_MODEL_TYPE, Type[QEFFBaseModel]] = {
-    QEFF_MODEL_TYPE.CAUSALLM: QEFFAutoModelForCausalLM
-}
-
-AUTO_MODEL_MAP_TO_MODEL_TYPE_MAP: Dict[Type[QEFFBaseModel], QEFF_MODEL_TYPE] = {v:k for k,v in MODEL_TYPE_TO_QEFF_AUTO_MODEL_MAP.items()}
-
-
-def get_hf_model_type(hf_model_path: str) -> QEFF_MODEL_TYPE:
-    """
-    Loads model config file and returns the type of the model (i.e. LLMs, SD, quantized etc.) as supported by the library.
-    """
-    assert os.path.isdir(hf_model_path), "Pleae pass local dir path where the model is downloaded; use `QEfficient.utils.login_and_download_hf_lm` for downloading hf model"
-    config, kwargs = AutoConfig.from_pretrained(
-                hf_model_path,
-                return_unused_kwargs=True,
-            )
-    
-    if config.__class__ in MODEL_FOR_CAUSAL_LM_MAPPING:
-        # FIXME: Add logic to handle if quantization config is stored in separate quant_config.json outside of config, also create a separate function for this and below lines
-        quant_config = getattr(config, "quantization_config", getattr(config, "quant_config", None))
-        if quant_config is not None:
-            if quant_config.get("quant_method", None) == "awq":
-                return QEFF_MODEL_TYPE.AWQ
-            else:
-                raise NotImplementedError(f"current model type is not yet supported {type(config)}")
-        else:
-            return QEFF_MODEL_TYPE.CAUSALLM
-    else:
-        raise NotImplementedError(f"model type {type(config)} is not yet supported")
